refactor(envs): split into seperate envs
